2004-02-20  graydon hoare  <graydon@pobox.com>
	* cert.cc, cert.hh, key.cc, key.hh, database.cc, database.hh:
	Add lots of little netsync support routines.
	* commands.cc (rebuild): Rehash everything too.
	* constants.cc (netcmd_minsz): Recalculate.
	* cryptopp/osrng.cpp (NonblockingRng::GenerateBlock): Handle
	/dev/urandom a bit better.
	* netcmd.cc, netcmd.hh: Remove describe cmds, add nonexistant cmd.
	* netio.hh: Add uleb128 stuff.
	* xdelta.cc: Add randomizing unit test suite.
	* diff_patch.cc: Remove commented-out dead line-merger code.
	* merkle_tree.cc: Fix various bugs.
	* netcmd.cc: Switch everything over to uleb128s.
	* netsync.cc: Implement lots of missing stuff.

+++ merkle_tree.hh
@@ -54,9 +54,9 @@ slot_state;
 
 struct merkle_node
 {    
-  u8 level;
+  size_t level;
   boost::dynamic_bitset<char> pref;
-  u64 total_num_leaves;
+  size_t total_num_leaves;
   boost::dynamic_bitset<char> bitmap;
   std::vector<id> slots;
   netcmd_item_type type;
